fix llvm issues in all windows num tests
