+++ crates/compiler/gen_llvm/src/llvm/build_str.rs
@@ -46,30 +46,6 @@ pub(crate) fn decode_from_utf8_result<'a, 'ctx, 'env>(
 }
 
 /// Dec.toStr : Dec -> Str
-pub(crate) fn dec_to_str<'a, 'ctx, 'env>(
-    env: &Env<'a, 'ctx, 'env>,
-    dec: BasicValueEnum<'ctx>,
-) -> BasicValueEnum<'ctx> {
-    let dec = dec.into_int_value();
-
-    let int_64 = env.context.i128_type().const_int(64, false);
-    let int_64_type = env.context.i64_type();
-
-    let dec_right_shift = env
-        .builder
-        .build_right_shift(dec, int_64, false, "dec_left_bits");
-
-    let right_bits = env.builder.build_int_cast(dec, int_64_type, "");
-    let left_bits = env.builder.build_int_cast(dec_right_shift, int_64_type, "");
-
-    call_str_bitcode_fn(
-        env,
-        &[],
-        &[right_bits.into(), left_bits.into()],
-        BitcodeReturns::Str,
-        bitcode::DEC_TO_STR,
-    )
-}
 
 /// Str.equal : Str, Str -> Bool
 pub(crate) fn str_equal<'a, 'ctx, 'env>(
